apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9
temporalio==1.6.0
openai==1.35.3
pyjwt==2.8.0
